+++ types/index.tst.ts
@@ -1,0 +1,18 @@
+import uri, {
+  type URIComponent,
+  type Options,
+} from '..'
+import { expect } from 'tstyche'
+
+const parsed = uri.parse('foo')
+expect(parsed).type.toBe<URIComponent>()
+
+const parsed2 = uri.parse('foo', {
+  domainHost: true,
+  scheme: 'https',
+  unicodeSupport: false
+})
+
+expect(parsed2).type.toBe<URIComponent>()
+expect({} as URIComponent).type.toBeAssignableTo<URIComponent>()
+expect({} as Options).type.toBeAssignableTo<Options>()
